% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beans.R
\docType{data}
\name{beans}
\alias{beans}
\title{Dry beans}
\source{
Koklu, M, and IA Ozkan. 2020. Multiclass Classification of Dry Beans Using
Computer Vision and Machine Learning Techniques.Computers and Electronics in
Agriculture, 174: 105507.
\doi{10.1016/j.compag.2020.105507},
\url{https://www.muratkoklu.com/datasets/}
}
\value{
\item{beans}{a tibble}
}
\description{
Dry beans
}
\details{
Koklu and Ozkan (2020) describe a data set where images of beans were
segmented and quantified. There are 13,611 beans quantified using 16
morphologic image features. The variety of each bean was labeled as one of
six values (in the column \code{class}).
}
\examples{
data(beans)
str(beans)
}
\keyword{datasets}
